backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.25.2

# AI Providers
google-generativeai==0.3.2
openai==1.6.1
anthropic==0.8.1
cohere==4.39

# HTTP & External APIs
httpx==0.25.2
aiohttp==3.9.1

# Utilities
redis==5.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
